requirements.txt
CairoSVG==2.6.0
Pillow==10.0.1
tqdm==4.67.1